(fix): Renamed "Aggregates" to "Services" for future MessageBus implementation inside.
