+++ Shop.Application/Contracts/Services/ICustomerService.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Shop.Application.Dto;
 
-namespace Shop.Application.Contracts.Aggregates;
+namespace Shop.Application.Contracts.Services;
 
-public interface ICustomerAggregate
+public interface ICustomerService
 {
     Task<IList<CustomerDto>> GetAllAsync();
     Task<CustomerDto> GetByIdAsync(Guid id);
