+++ Shop.Application/Contracts/Services/IOrderService.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Shop.Application.Dto;
 
-namespace Shop.Application.Contracts.Aggregates;
+namespace Shop.Application.Contracts.Services;
 
-public interface IOrderAggregate
+public interface IOrderService
 {
     Task<IList<OrderDto>> GetAllAsync();
     Task<OrderDto> GetByIdAsync(Guid id);
